+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Check PVDB history and ensure that each trading day has a value for all FRED measurements.
+
+### Changed
+- Always set split factor in quote to 1
+
+### Removed
 
 ## [0.1.0] - 2022-05-20
 ### Added